Resolving sonar devon architecture issues
