+++ java/mtsj/api/src/main/java/com/devonfw/application/mtsj/bookingmanagement/common/api/to/InvitedGuestEto.java
@@ -1,4 +1,4 @@
-package com.devonfw.application.mtsj.bookingmanagement.logic.api.to;
+package com.devonfw.application.mtsj.bookingmanagement.common.api.to;
 
 import java.sql.Timestamp;
