+++ java/mtsj/api/src/main/java/com/devonfw/application/mtsj/usermanagement/common/api/to/UserCto.java
@@ -1,9 +1,9 @@
-package com.devonfw.application.mtsj.usermanagement.logic.api.to;
+package com.devonfw.application.mtsj.usermanagement.common.api.to;
 
 import java.util.List;
 
-import com.devonfw.application.mtsj.bookingmanagement.logic.api.to.BookingEto;
-import com.devonfw.application.mtsj.dishmanagement.logic.api.to.DishEto;
+import com.devonfw.application.mtsj.bookingmanagement.common.api.to.BookingEto;
+import com.devonfw.application.mtsj.dishmanagement.common.api.to.DishEto;
 import com.devonfw.module.basic.common.api.to.AbstractCto;
 
 /**
