+++ java/mtsj/api/src/main/java/com/devonfw/application/mtsj/usermanagement/logic/api/Usermanagement.java
@@ -2,10 +2,10 @@ package com.devonfw.application.mtsj.usermanagement.logic.api;
 
 import org.springframework.data.domain.Page;
 
+import com.devonfw.application.mtsj.usermanagement.common.api.to.UserEto;
+import com.devonfw.application.mtsj.usermanagement.common.api.to.UserRoleEto;
 import com.devonfw.application.mtsj.usermanagement.common.api.to.UserRoleSearchCriteriaTo;
 import com.devonfw.application.mtsj.usermanagement.common.api.to.UserSearchCriteriaTo;
-import com.devonfw.application.mtsj.usermanagement.logic.api.to.UserEto;
-import com.devonfw.application.mtsj.usermanagement.logic.api.to.UserRoleEto;
 
 /**
  * Interface for Usermanagement component.
