+++ java/mtsj/core/src/main/java/com/devonfw/application/mtsj/dishmanagement/logic/impl/DishmanagementImpl.java
@@ -16,8 +16,12 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 
 import com.devonfw.application.mtsj.dishmanagement.common.api.Category;
+import com.devonfw.application.mtsj.dishmanagement.common.api.to.CategoryEto;
 import com.devonfw.application.mtsj.dishmanagement.common.api.to.CategorySearchCriteriaTo;
+import com.devonfw.application.mtsj.dishmanagement.common.api.to.DishCto;
+import com.devonfw.application.mtsj.dishmanagement.common.api.to.DishEto;
 import com.devonfw.application.mtsj.dishmanagement.common.api.to.DishSearchCriteriaTo;
+import com.devonfw.application.mtsj.dishmanagement.common.api.to.IngredientEto;
 import com.devonfw.application.mtsj.dishmanagement.common.api.to.IngredientSearchCriteriaTo;
 import com.devonfw.application.mtsj.dishmanagement.dataaccess.api.CategoryEntity;
 import com.devonfw.application.mtsj.dishmanagement.dataaccess.api.DishEntity;
@@ -26,12 +30,8 @@ import com.devonfw.application.mtsj.dishmanagement.dataaccess.api.repo.CategoryR
 import com.devonfw.application.mtsj.dishmanagement.dataaccess.api.repo.DishRepository;
 import com.devonfw.application.mtsj.dishmanagement.dataaccess.api.repo.IngredientRepository;
 import com.devonfw.application.mtsj.dishmanagement.logic.api.Dishmanagement;
-import com.devonfw.application.mtsj.dishmanagement.logic.api.to.CategoryEto;
-import com.devonfw.application.mtsj.dishmanagement.logic.api.to.DishCto;
-import com.devonfw.application.mtsj.dishmanagement.logic.api.to.DishEto;
-import com.devonfw.application.mtsj.dishmanagement.logic.api.to.IngredientEto;
 import com.devonfw.application.mtsj.general.logic.base.AbstractComponentFacade;
-import com.devonfw.application.mtsj.imagemanagement.logic.api.to.ImageEto;
+import com.devonfw.application.mtsj.imagemanagement.common.api.to.ImageEto;
 
 /**
  * Implementation of component interface of dishmanagement
