+++ java/mtsj/core/src/main/java/com/devonfw/application/mtsj/dishmanagement/service/api/rest/DishmanagementRestService.java
@@ -12,14 +12,14 @@ import javax.ws.rs.core.MediaType;
 import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.CrossOrigin;
 
+import com.devonfw.application.mtsj.dishmanagement.common.api.to.CategoryEto;
 import com.devonfw.application.mtsj.dishmanagement.common.api.to.CategorySearchCriteriaTo;
+import com.devonfw.application.mtsj.dishmanagement.common.api.to.DishCto;
+import com.devonfw.application.mtsj.dishmanagement.common.api.to.DishEto;
 import com.devonfw.application.mtsj.dishmanagement.common.api.to.DishSearchCriteriaTo;
+import com.devonfw.application.mtsj.dishmanagement.common.api.to.IngredientEto;
 import com.devonfw.application.mtsj.dishmanagement.common.api.to.IngredientSearchCriteriaTo;
 import com.devonfw.application.mtsj.dishmanagement.logic.api.Dishmanagement;
-import com.devonfw.application.mtsj.dishmanagement.logic.api.to.CategoryEto;
-import com.devonfw.application.mtsj.dishmanagement.logic.api.to.DishCto;
-import com.devonfw.application.mtsj.dishmanagement.logic.api.to.DishEto;
-import com.devonfw.application.mtsj.dishmanagement.logic.api.to.IngredientEto;
 
 /**
  * The service interface for REST calls in order to execute the logic of component {@link Dishmanagement}.
