+++ java/mtsj/core/src/main/java/com/devonfw/application/mtsj/imagemanagement/service/impl/rest/ImagemanagementRestServiceImpl.java
@@ -5,9 +5,9 @@ import javax.inject.Named;
 
 import org.springframework.data.domain.Page;
 
+import com.devonfw.application.mtsj.imagemanagement.common.api.to.ImageEto;
 import com.devonfw.application.mtsj.imagemanagement.common.api.to.ImageSearchCriteriaTo;
 import com.devonfw.application.mtsj.imagemanagement.logic.api.Imagemanagement;
-import com.devonfw.application.mtsj.imagemanagement.logic.api.to.ImageEto;
 import com.devonfw.application.mtsj.imagemanagement.service.api.rest.ImagemanagementRestService;
 
 /**
