+++ java/mtsj/core/src/main/java/com/devonfw/application/mtsj/ordermanagement/common/api/exception/NoInviteException.java
@@ -4,7 +4,7 @@ import net.sf.mmm.util.exception.api.NlsRuntimeException;
 
 /**
  * This exception is thrown if the guest token of an
- * {@link com.devonfw.application.mtsj.ordermanagement.logic.api.to.OrderEto} has no
+ * {@link com.devonfw.application.mtsj.ordermanagement.common.api.to.OrderEto} has no
  * {@link com.devonfw.application.mtsj.bookingmanagement.dataaccess.api.BookingEntity} related.
  *
  */
