+++ java/mtsj/core/src/main/java/com/devonfw/application/mtsj/usermanagement/service/api/rest/UsermanagementRestService.java
@@ -12,10 +12,10 @@ import javax.ws.rs.core.MediaType;
 import org.springframework.data.domain.Page;
 
 import com.devonfw.application.mtsj.general.common.api.Usermanagement;
+import com.devonfw.application.mtsj.usermanagement.common.api.to.UserEto;
+import com.devonfw.application.mtsj.usermanagement.common.api.to.UserRoleEto;
 import com.devonfw.application.mtsj.usermanagement.common.api.to.UserRoleSearchCriteriaTo;
 import com.devonfw.application.mtsj.usermanagement.common.api.to.UserSearchCriteriaTo;
-import com.devonfw.application.mtsj.usermanagement.logic.api.to.UserEto;
-import com.devonfw.application.mtsj.usermanagement.logic.api.to.UserRoleEto;
 
 /**
  * The service interface for REST calls in order to execute the logic of component {@link Usermanagement}.
