+++ java/mtsj/core/src/main/java/com/devonfw/application/mtsj/usermanagement/service/impl/rest/UsermanagementRestServiceImpl.java
@@ -5,11 +5,11 @@ import javax.inject.Named;
 
 import org.springframework.data.domain.Page;
 
+import com.devonfw.application.mtsj.usermanagement.common.api.to.UserEto;
+import com.devonfw.application.mtsj.usermanagement.common.api.to.UserRoleEto;
 import com.devonfw.application.mtsj.usermanagement.common.api.to.UserRoleSearchCriteriaTo;
 import com.devonfw.application.mtsj.usermanagement.common.api.to.UserSearchCriteriaTo;
 import com.devonfw.application.mtsj.usermanagement.logic.api.Usermanagement;
-import com.devonfw.application.mtsj.usermanagement.logic.api.to.UserEto;
-import com.devonfw.application.mtsj.usermanagement.logic.api.to.UserRoleEto;
 import com.devonfw.application.mtsj.usermanagement.service.api.rest.UsermanagementRestService;
 
 /**
