+++ java/mtsj/core/src/test/java/com/devonfw/application/mtsj/dishmanagement/logic/impl/DishmanagementTest.java
@@ -13,10 +13,10 @@ import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Direction;
 
 import com.devonfw.application.mtsj.SpringBootApp;
+import com.devonfw.application.mtsj.dishmanagement.common.api.to.CategoryEto;
+import com.devonfw.application.mtsj.dishmanagement.common.api.to.DishCto;
 import com.devonfw.application.mtsj.dishmanagement.common.api.to.DishSearchCriteriaTo;
 import com.devonfw.application.mtsj.dishmanagement.logic.api.Dishmanagement;
-import com.devonfw.application.mtsj.dishmanagement.logic.api.to.CategoryEto;
-import com.devonfw.application.mtsj.dishmanagement.logic.api.to.DishCto;
 import com.devonfw.module.test.common.base.ComponentTest;
 
 /**
